rankedNN hot-fix
Errors were emerging from plotting the original origin stomata coordinates for a given FOV, in part due to implementation of the ranking method by means of numpy's argsort which had undesired effects. Have since restructured this function to allow better assessment of the raw data used by the KDD analyses.  In addition, have built a plot_rankedNN() function to allow for these edges of NN affiliation to be seen as an illustration.