Rorshach and Peak_spacing function reimplementation
Within the rorshach function switch from a for looping function to iterate through each kernel to now split the KDD into quadrants which are then inverted using matrix functions from numpy greatly increasing speed and consistency of the normalization procedure.  Also rebuilt peak_spacing where the buffer variable (i.e. the stomatal width) is used to specify some minimal barrier between the left and right planes of the in-file masking given that Sorghum WEST variances are high enough that splitting on both the trench and across the y-origin is now warranted.